# frontend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations": {
      "post": {
        "operationId": "OrganizationController_createOrganization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDtoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationSummaryDtoResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create organization",
        "tags": ["Organizations"]
      },
      "get": {
        "operationId": "OrganizationController_getMyOrganizations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationSummaryDtoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get organizations for current user",
        "tags": ["Organizations"]
      }
    },
    "/organizations/personal": {
      "post": {
        "operationId": "OrganizationController_createPersonalWorkspace",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationSummaryDtoResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create personal workspace",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}": {
      "get": {
        "operationId": "OrganizationController_getOrganizationDetails",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetailsDtoResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get organization details",
        "tags": ["Organizations"]
      },
      "patch": {
        "operationId": "OrganizationController_editOrganization",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDtoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDtoResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Edit organization",
        "tags": ["Organizations"]
      },
      "delete": {
        "operationId": "OrganizationController_softDeleteOrganization",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Soft delete organization",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/transfer-ownership": {
      "post": {
        "operationId": "OrganizationController_transferOwnership",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipDtoRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Transfer organization ownership",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/invite": {
      "post": {
        "operationId": "OrganizationController_inviteMember",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDtoRequest"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Invite member to organization",
        "tags": ["Organizations"]
      }
    },
    "/organizations/accept-invite": {
      "post": {
        "operationId": "OrganizationController_acceptInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteDtoRequest"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Accept organization invite",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/members/{userId}": {
      "delete": {
        "operationId": "OrganizationController_removeMember",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Remove member from organization",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/leave": {
      "post": {
        "operationId": "OrganizationController_leaveOrganization",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Leave organization",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/members/{userId}/role": {
      "patch": {
        "operationId": "OrganizationController_changeMemberRole",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMemberRoleDtoRequest"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Change member role",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/members/{userId}/suspend": {
      "patch": {
        "operationId": "OrganizationController_suspendMember",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Suspend member",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/members/{userId}/reactivate": {
      "patch": {
        "operationId": "OrganizationController_reactivateMember",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Reactivate member",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/members": {
      "get": {
        "operationId": "OrganizationController_listMembers",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationMemberDtoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "List organization members",
        "tags": ["Organizations"]
      }
    },
    "/organizations/{orgId}/users/search": {
      "get": {
        "operationId": "OrganizationController_searchUsers",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": { "example": "john", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationUserSearchDtoResponse"
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Search existing users for invitation",
        "tags": ["Organizations"]
      }
    },
...
  },
  "info": {
    "title": "Survix API",
    "description": "Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "CreateOrganizationDtoRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Acme Inc" },
          "slug": { "type": "string", "example": "acme-inc" },
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "PRIVATE"],
            "example": "PRIVATE"
          },
          "description": { "type": "string" },
          "industry": { "type": "string" },
          "size": { "type": "string" },
          "websiteUrl": { "type": "string" },
          "contactEmail": { "type": "string" }
        },
        "required": ["name", "slug", "visibility"]
      },
      "OrganizationSummaryDtoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "60f7f0b3b4d1c2a5f8e9a123" },
          "name": { "type": "string", "example": "Acme Inc" },
          "slug": { "type": "string", "example": "acme-inc" },
          "role": { "type": "string", "enum": ["OWNER", "ADMIN", "MEMBER"] },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "ARCHIVED", "SUSPENDED"]
          },
          "accountType": {
            "type": "string",
            "enum": ["PERSONAL", "ORGANIZATION"]
          },
          "isPersonal": { "type": "boolean" }
        },
        "required": [
          "id",
          "name",
          "slug",
          "role",
          "status",
          "accountType",
          "isPersonal"
        ]
      },
      "OrganizationDtoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "logoUrl": { "type": "object" },
          "ownerId": { "type": "string" },
          "accountType": {
            "type": "string",
            "enum": ["PERSONAL", "ORGANIZATION"]
          },
          "isPersonal": { "type": "boolean" },
          "description": { "type": "object" },
          "industry": { "type": "object" },
          "size": { "type": "object" },
          "websiteUrl": { "type": "object" },
          "contactEmail": { "type": "object" },
          "visibility": { "type": "string", "enum": ["PUBLIC", "PRIVATE"] },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "ARCHIVED", "SUSPENDED"]
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "ownerId",
          "accountType",
          "isPersonal",
          "visibility",
          "status"
        ]
      },
      "OrganizationDetailsDtoResponse": {
        "type": "object",
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationDtoResponse"
          },
          "currentUserRole": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "MEMBER"]
          },
          "memberStatus": {
            "type": "string",
            "enum": ["ACTIVE", "INVITED", "SUSPENDED", "LEFT"]
          }
        },
        "required": ["organization", "currentUserRole", "memberStatus"]
      },
      "UpdateOrganizationDtoRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Acme Inc" },
          "description": { "type": "string" },
          "industry": { "type": "string" },
          "size": { "type": "string" },
          "websiteUrl": { "type": "string" },
          "contactEmail": { "type": "string" },
          "visibility": { "type": "string", "enum": ["PUBLIC", "PRIVATE"] }
        }
      },
      "TransferOwnershipDtoRequest": {
        "type": "object",
        "properties": {
          "newOwnerUserId": {
            "type": "string",
            "description": "New owner user ID (Mongo ObjectId)"
          }
        },
        "required": ["newOwnerUserId"]
      },
      "InviteMemberDtoRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@gmail.com" },
          "role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "MEMBER"],
            "example": "MEMBER"
          }
        },
        "required": ["email", "role"]
      },
      "AcceptInviteDtoRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Raw invite token sent via email"
          }
        },
        "required": ["token"]
      },
      "ChangeMemberRoleDtoRequest": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "MEMBER"],
            "example": "ADMIN"
          }
        },
        "required": ["role"]
      },
      "OrganizationMemberUserDtoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "username": { "type": "object" },
          "name": { "type": "object" },
          "avatar": { "type": "object" }
        },
        "required": ["id", "email"]
      },
      "OrganizationMemberDtoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "organizationId": { "type": "string" },
          "role": { "type": "string", "enum": ["OWNER", "ADMIN", "MEMBER"] },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INVITED", "SUSPENDED", "LEFT"]
          },
          "joinedAt": { "type": "object" },
          "leftAt": { "type": "object" },
          "user": {
            "$ref": "#/components/schemas/OrganizationMemberUserDtoResponse"
          }
        },
        "required": ["id", "userId", "organizationId", "role", "status"]
      },
      "OrganizationUserSearchDtoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "username": { "type": "object" },
          "name": { "type": "object" },
          "avatar": { "type": "object" }
        },
        "required": ["id", "email"]
      },
...
    }
  }
}